Update deployment suffix generation
Make the suffix somewhat unique with hash(gitsha, timestamp){,7}
to make the rollouts idempotent.

This fix mitigates issues when a workflow is manually triggered
and generating the same suffix for a deployment - which in turn
issues a cleanup process for the same deploy -> no active deploys/pods
available in cluster.